open_aoi_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "open_aoi"
version="0.0.14"
authors = [
  { name="Cherniaev Egor", email="chrnyaevek@gmail.com" },
]
description = "Support python package for Open AOI system project."
keywords=["automated optical inspection", "pcb", "computer vision"]
readme = "README.md"
requires-python = ">=3.7"
dependencies = ["python-dotenv", "opencv-python", "nicegui", "sqlalchemy", "pymysql", "cryptography", "bcrypt", "pillow", "minio"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/ChrnyaevEK/open-aoi"
Issues = "https://github.com/ChrnyaevEK/open-aoi/issues"